requirements.txt
flask==3.0.3
gunicorn==22.0.0
requests==2.31.0
cryptography==42.0.5
colorama==0.4.6
tabulate==0.9.0
python-dotenv==1.0.1
